+++ client/src/Home.js
@@ -1,0 +1,5 @@
+function Home() {
+    return <h1>HOME TEST</h1>;
+}
+
+export default Home;
